reverted the prices in double from integer and updated the test

+++ src/main/java/com/hmrc/shoppping/basket/BasketPriceCalculator.java
@@ -16,42 +16,42 @@ import rx.Observable;
 public class BasketPriceCalculator {
 
 	private Observable<String> basketObservable;
-	private Observable<Integer> totalPriceObservable;
-	private Observable<Integer> totalPromotionalPriceObservable;
+	private Observable<Double> totalPriceObservable;
+	private Observable<Double> totalPromotionalPriceObservable;
 
 	public BasketPriceCalculator(List<String> basketItems) {
 		basketObservable = Observable.from(basketItems);
-		totalPriceObservable = Observable.merge(getApplePrice(), getOrangePrice()).scan(0, this::sum);
+		totalPriceObservable = Observable.merge(getApplePrice(), getOrangePrice()).scan(0.0, this::sum);
 		totalPromotionalPriceObservable = Observable.merge(getPromotionalApplePrice(), getPromotionalOrangePrice())
-				.scan(0, this::sum);
+				.scan(0.0, this::sum);
 	}
 
-	public Observable<Integer> totalActualPrice() {
+	public Observable<Double> totalActualPrice() {
 		return totalPriceObservable;
 	}
 
-	public Observable<Integer> totalPromotionalActualPrice() {
+	public Observable<Double> totalPromotionalActualPrice() {
 		return totalPromotionalPriceObservable;
 	}
 
-	private Observable<Integer> getApplePrice() {
+	private Observable<Double> getApplePrice() {
 		return basketObservable.filter(ItemEnum.APPLE.name()::equals).map((eachOffer) -> PriceConstants.APPLE_RATE);
 	}
 
-	private Observable<Integer> getOrangePrice() {
+	private Observable<Double> getOrangePrice() {
 		return basketObservable.filter(ItemEnum.ORANGE.name()::equals).map((eachOffer) -> PriceConstants.ORANGE_RATE);
 	}
 
-	private Observable<Integer> getPromotionalApplePrice() {
+	private Observable<Double> getPromotionalApplePrice() {
 		return basketObservable.filter(ItemEnum.APPLE.name()::equals).buffer(2).map((eachOffer) -> PriceConstants.APPLE_RATE);
 	}
 
-	private Observable<Integer> getPromotionalOrangePrice() {
+	private Observable<Double> getPromotionalOrangePrice() {
 		return basketObservable.filter(ItemEnum.ORANGE.name()::equals).buffer(3)
 				.map((eachOffer) -> eachOffer.size() == 3 ? PriceConstants.ORANGE_RATE * 2 : PriceConstants.ORANGE_RATE * eachOffer.size());
 	}
 
-	private Integer sum(Integer acc, Integer next) {
+	private Double sum(Double acc, Double next) {
 		return acc + next;
 	}
 }
